fixed physical gate coordinates and corresponding functions

+++ include/hal_core/netlist/gate.h
@@ -102,7 +102,7 @@ namespace hal
          *
          * @param[in] x - The gate's x-coordinate.
          */
-        void set_location_x(float x);
+        void set_location_x(i32 x);
 
         /**
          * Get the physical location x-coordinate of the gate in the layout.<br>
@@ -110,7 +110,7 @@ namespace hal
          *
          * @returns The gate's x-coordinate.
          */
-        float get_location_x() const;
+        i32 get_location_x() const;
 
         /**
          * Set the physical location y-coordinate of the gate in the layout.<br>
@@ -118,7 +118,7 @@ namespace hal
          *
          * @param[in] y - The gate's y-coordinate.
          */
-        void set_location_y(float y);
+        void set_location_y(i32 y);
 
         /**
          * Get the physical location y-coordinate of the gate in the layout.<br>
@@ -126,7 +126,7 @@ namespace hal
          *
          * @returns The gate's y-coordinate.
          */
-        float get_location_y() const;
+        i32 get_location_y() const;
 
         /**
          * Set the physical location of the gate in the layout.<br>
@@ -134,7 +134,7 @@ namespace hal
          *
          * @param[in] location - A pair <x-coordinate, y-coordinate>.
          */
-        void set_location(const std::pair<float, float>& location);
+        void set_location(const std::pair<i32, i32>& location);
 
         /**
          * Get the physical location of the gate in the layout.<br>
@@ -143,7 +143,7 @@ namespace hal
          *
          * @returns A pair <x-coordinate, y-coordinate>.
          */
-        std::pair<float, float> get_location() const;
+        std::pair<i32, i32> get_location() const;
 
         /**
          * Get the module which contains this gate.
